third_party/WebKit/LayoutTests/media/track/opera/interfaces/TextTrackCue/text-expected.txt

 
PASS TextTrackCue.text, script-created cue 
FAIL TextTrackCue.text, parsed cue assert_equals: expected "" but got "test"
